+++ src/Header/Request/IfMatch.php
@@ -41,7 +41,7 @@ final class IfMatch implements Header
      */
     public static function fromValue(string $value): self
     {
-        $regEx = '{^' . Rfc7230::hashRule(Rfc7232::ENTITY_TAG, 1) . '$}';
+        $regEx = '{^(?:\\*|' . Rfc7230::hashRule(Rfc7232::ENTITY_TAG, 1) . ')$}';
         if (utf8_decode($value) !== $value || preg_match($regEx, $value) !== 1) {
             throw new InvalidHeader('Invalid header value: ' . $value);
         }