+++ tests/Header/Request/IfNoneMatchTest.php
@@ -50,6 +50,17 @@ final class IfNoneMatchTest extends TestCase
     }
 
     /**
+     * @covers ::fromValue
+     */
+    public function testCanConstructIfNoneMatchWithoutEntityTagsFromValue(): void
+    {
+        $ifNoneMatch = new IfNoneMatch(new EntityTagSet());
+        $ifNoneMatchFromValue = IfNoneMatch::fromValue('*');
+
+        self::assertEquals($ifNoneMatch, $ifNoneMatchFromValue);
+    }
+
+    /**
      * @covers ::__toString
      */
     public function testCanConvertIfNoneMatchWithoutEntityTagsToString(): void